Security: checks and automatic workflow (#41)
* fix: remove ConnectionManager, fix mutex deadlocks

* feat: redesign config editor UI and fix edge cases

* fix: sync connectivity test to "Save & Test" button

* feat: refactor Commanding Panel and add layout options

* fix: bug with Commanding Panel layout

* feat: querying commands from panel directly in Commanding panel

* fix: refactor QueryEditor and components code

* feat: rewrite Links panels, make Command History and Alarms panel not use interval time

* chore: add PROVISIONING.md and add datasource provisioned configuration

* feat: add time sync using panel and dashboard time injection

* BREAKING CHANGE: update to latest SDK and Grafana 13

* fix: backend hot reloading for dev environement

* fix: commanding button input bug

* fix: issues with time sync panel

* chore: document time sync panel, setup instructions with backend watcher, and some visuals

* chore: lint

* security: fix pnpm audit

* chore: fix pnpm unmet peer deps

* security: sanitize input in ImageRenderer

* security: fix DoS vulnerabilities

* security: address known go CVEs by updating to 1.26.4

* security: setup weekly dependabot checks and security audits with PR/issue opens

* fix: dependency version issue

* security: add PR-triggered gosec and validator security checks

* fix: int32 conversion, pnpm workflow setup

* fix: remove version from pnpm action

* feat: add verbose and human-readable PR comments for both audit and checks

* fix: plugin-validator json output

* fix: backticks in yaml

* CI: disable analyzers and add nested plugin

* CI: delete security comment on success

+++ pkg/yamcs/client/command_history_subscription.go
@@ -13,7 +13,7 @@ type CommandHistoryListener func(entry *commanding.CommandHistoryEntry)
 
 // CommandHistorySubscription manages a subscription to command history updates.
 type CommandHistorySubscription struct {
-	subscriptionID      int
+	subscriptionID      int32
 	activeSubscriptions types.Set[string]
 	commandListener     CommandHistoryListener
 	Instance            string
@@ -75,7 +75,7 @@ func (client *YamcsClient) HandleCommandMessage(message *api.ServerMessage) {
 		}
 
 		callID := message.GetCall()
-		subscription, found := client.CommandHistorySubscriptions[int(callID)]
+		subscription, found := client.CommandHistorySubscriptions[callID]
 		if found && subscription.commandListener != nil {
 			subscription.commandListener(entry)
 		}
@@ -93,7 +93,7 @@ func (subscription *CommandHistorySubscription) Halt() {
 	delete(subscription.client.CommandHistorySubscriptions, subscription.subscriptionID)
 
 	cancelRequest := &api.CancelOptions{
-		Call: int32(subscription.subscriptionID),
+		Call: subscription.subscriptionID,
 	}
 
 	anyMessage, _ := anypb.New(cancelRequest)
